requirements.txt
pyyaml==5.4.1
tqdm==4.59.0
konlpy==0.6.0
mecab-python3==1.0.5
